move components to ViewForm folder to share styles
