+++ src/shared/components/ViewForm/ElasticSearchQueryForm.tsx
@@ -7,7 +7,7 @@ import 'codemirror/addon/display/placeholder';
 import 'codemirror/mode/javascript/javascript';
 
 // TODO move once SparqlQueryView is also refactored
-import '../RawQueryView/view-form.less';
+import './view-form.less';
 
 const FormItem = Form.Item;
 const ListItem = List.Item;
