+++ src/shared/components/ViewForm/view-form.less
@@ -10,6 +10,12 @@
     overflow: 'scroll';
     margin-bottom: @default-pad;
   }
+  .ant-card-body {
+    overflow: scroll;
+  }
+  .results .empty {
+    color: @error-color;
+  }
   .react-json-view {
     overflow-x: scroll;
   }
